Rename kafka modules to event_sub and event_pub
